Move code out of Startup
Move the puzzles API into its own class.
